Put a default microblaze elf into the bitstream.
A prebuilt elf file is associated with the microblaze in the IPI TCL, this
gets it into the bitstream (.rcdo file).

The pdi file generated in the bootgen directory brings the microblaze out of reset
during boot by including the microblaze_reset.cdo file which is generated from
the microblaze_reset.txt file.