[air] Always cross-compile 'libairhost'
The runtime libraries are odd because they should always be compiled
for the the target architecture (probably arm, but might be microblaze
at some point).  This patch factors out the air/runtime_lib directory
(consisting primarily of libairhost) into a separate external project
driven from the toplevel, enabling herds to be completely compiled
both in an x86 environment (with cross-compilation) or in a target
environment.