Add aircc.py to compile AIR dialect programs
This tool runs the compilation steps for each herd in an AIR
program. Herds are lowered to AIE dialect and aiecc.py is invoked
to generate ELF files for each core in each herd. The host side
configuration and control code generated by aiecc.py and aircc.py
is packaged into a single static library with herd information
exported to the AIR runtime as externally visible symbols.