Refactor pipeline lowering code
This change refactors pipeline lowering in air-to-aie pass by
by lowering air.pipeline to a series of affine.if ops containing
air.pipeline.{get,put} ops. This is expsed to the command line
as -air-pipeline-to-affine.

The air-to-aie pass has been updated to handle affine.if as a more
generic AIE.core specialization mechanism.