Updates to env_setup.sh
This script is intended to be sourced, not executed. Therefore, it should not
have the executable attribute in the file system and not start with the #!
for executing scripts.

If a user is connected over ssh and does not provide the correct number of
parameters while sourcing this script, 'exit' will close the ssh connection
which is quite a harsh punishment. Use 'return' instead to give the user
another chance.

Signed-off-by: Joel Nider <joel.nider@xilinx.com>