Save environment to an external file
The environment (paths to build tools) has to be set up for each shell.
That means typing the paths to the MLIR_AIE MLIR_AIR and LLVM each time.
These variables are unlikely to change very quickly, so it can save the
user a lot of time if they are set once and saved to a file that can be
reloaded again later.

Signed-off-by: Joel Nider <joel.nider@xilinx.com>