controller: Use macros for firmware version
The firmware version is dependent on the architecture. In a Microblaze
system, the version comes from the first controller. In an ARM system,
no version number is printed.

Following the Microblaze encoding of the version number, add macros to
encode and decode a version number. Apply those macros to the ARM
version as well.

Also add REG_HERD_CONTROLLER_COUNT to describe the offset rather than
a literal.

Signed-off-by: Joel Nider <joel.nider@xilinx.com>