controller: keep copy of ARM and microblaze BSP
The BSP (board support package) is a kind of simple operating system that
provides basic functionality for Xilinx hardware. Vitis tools compile this
BSP automatically and place a copy of the source in the project directory
for potential modifications. Until now, we have been creating the project
fresh each time, which copies the BSP from a local Vitis installation. By
storing an explicit copy, we can accomplish a few goals:
1) Break dependence on Vitis: we no longer need to have a local Vitis
   installed to compile this firmware. We still use the compiler from Vitis
	but this is more from convenience than necessity.
2) Improve build times: the BSP rarely changes (if ever) so compiling it each
	time accomplishes nothing. Keeping a static copy brings the build time
	down from 1m30 to 10s.
3) Enables a unified Makefile for multiple platforms: we can now perform
	conditional compilation with a single Makefile when targeting different
	platforms that share the same source tree.

Signed-off-by: Joel Nider <joel.nider@xilinx.com>