controller: remove dead code
This code is never called so it is just increasing the size of the executable
and generating warnings. It looks like some of this is for debugging and some
is for future functionality. When I find out exactly what the functions are
for, I will put them back with conditional compilation flags.

Signed-off-by: Joel Nider <joel.nider@xilinx.com>