Update links to .md files
Some of the links appear broken when browsing the source repo directly.
They point to .html files that are generated when the docs are
published to github.io and do not exist in the source repo. The links
would need to point to .md files in source repo and .html files in
the published docs to be correct in both cases.

Github has a script that rewrites links to .md files to .html
automatically upon publishing. Therefore, the correct form is to link
to .md files, as explained in this post:
https://github.blog/2016-12-05-relative-links-for-github-pages/

Change the links from .html to .md so they can be browsed in the
repo.