driver: create sysfs interface
Use sysfs files to access the AIE address space. The address space is
exposed over a BAR by the device but only the driver has access to it
by default. Any user code must access the address space through the
sysfs files; address and value.

To bypass this sysfs mechanism, the driver can still be loaded with
the parameter 'enable_aie=1'. This will enable legacy apps to mmap
the AIE BAR directly.

This requires an updated version of libxaie (xaienginev2 driver) that
contains the sysfs backend and ability to select the backend at
runtime.
https://github.com/jnider/embeddedsw/tree/vck5000

Signed-off-by: Joel Nider <joel@ece.ubc.ca>