driver: update license to GPL-2.0
All Linux kernel code must be under a GPL v2 license. In case this
driver is ever added to the upstream kernel, it is a good idea if
the license is set from the beginning.

Setting the license to MIT gives the following error:

"FATAL: modpost: GPL-incompatible module amdair.ko uses GPL-only
symbol 'device_destroy'"

Signed-off-by: Joel Nider <joel.nider@amd.com>