driver: add multi-queue support
Multiple contexts can share a single AIE device by using separate command
queues. In this first implementation, each command processor exposes a
single queue. On a BP-based platform, that means 7 queues. There must be
a mechanism to allocate and free these queues for use by separate contexts.

In the first case, queue 0 (belonging to the ARM processor) is reserved for
administration functions that can be performed only by the driver. This
includes the read32/write32 commands and possibly global administration
functions (such as card-level reset and initialize) in the future.

The mmap() interface has changed to support multiple devices. The offset
field is now encoded using a macro found in amdair_ioctl.h. The macro
encodes the target device, the memory range (BRAM, DRAM, etc.) the offset
inside that range.

The IOCTLs have been renamed to move away from the KFD starting point.

In the future, we must also add a mechanism to prevent mapping memory for a
queue that has not been created or not owned by the mapping process.

Signed-off-by: Joel Nider <joel.nider@amd.com>