test: add dependency on libelf
The runtime lib was modified to use libelf for loading airbin files. Since the
runtime lib is built as a static library (.a file), it does not include a link
step and can therefore not be linked directly with libelf. However, it does
call functions in libelf when loading airbin files, which means it has a
dependency on libelf. That forces all applications that link with the runtime
lib to also link with libelf, even if those functions are not used in that
particular application.

The command line options are also put on separate lines and sorted by type
(library path, include path, etc.) to make it easier to modify, diff and
understand what is happening in general.

Signed-off-by: Joel Nider <joel.nider@amd.com>