runtime: Load airbin to device
An AIRBIN is an ELF file that contains AIE configuration and executable code.
It is loaded to the device by the runtime by copying it to device memory and
then notifying the device that continues the process.

Add code to read the file, place it in device memory and notify the device.

This depends on libelf (from elfutils) that is built in:
github-clone-build-elfutils.sh

Signed-off-by: Joel Nider <joel.nider@amd.com>