controller: add firmware support for airbin
Add a new packet type AIR_PKT_TYPE_AIRBIN for loading an AIRBIN. The packet
holds the base address in device memory where the AIRBIN has been loaded by
the runtime_lib on the host. The packet handler builds a CDMA descriptor
chain based on the AIRBIN sections and loads them into AIE memory.

The ARM_CONTROLLER define is set by the compiler so it is available to all
source files.

The CDMA driver is moved to a separate source file. Currently, only the
scatter/gather mode has been implemented. A future patch will add the simple
mode as well, and then the legacy CDMA code can be removed.

Add a function to reset a shim tile. Follows the procedure for shim tile resets.
The reset is a global action but masked by registers belonging to each shim
tile. By unmasking a particular tile, it will participate in the reset. The
actual reset is performed through the NPI interface, then all the registers are
unmasked to prevent accidental reset later.

Signed-off-by: Joel Nider <joel.nider@amd.com>