utils: use a parameter name for git clone
When cloning sub-projects, use a parameter name for the target directory. This
makes the script a bit easier to read, and easier to check for existing
directories before cloning. This does not change the behaviour at all, but it
does suppress a non-fatal error when trying to clone to an existing directory.

Signed-off-by: Joel Nider <joel.nider@amd.com>