Add basic ipu xclbin support to aircc (#454)
If device == "ipu" in aircc, then lower airrt to ipu instruction sequence instead of air runtime host code, and request xclbin from aiecc instead of elfs.

air-insert-launch-and-segment-around-herd, air-dma-to-channel, and air-specialize-channel-wrap-and-stride are added to the standard aircc pipeline. air-dma-to-channel is enabled only for device="ipu". It is not safe to run for airhost (aie1) target because the runtime cannot handle the generated dma schedule.

Tests xrt/02_mul_shim_i32 and xrt/03_mul_shim_i16 are ported to use aircc to compile air dialect to xclbin.