Replace affine load and store with memref variants
WIP: First steps toward build/run on npu for python example. Builds but does not run yet.

test.ccp is not correct but it is now running on the device. Also removed lit files for now while I debug

wip: small progress

test is running now, but does not check values in a meaningful way

'make run' now passes with a meaningful check!